break out WorkspaceSettingRepository
